meta:lib:oeqa:utils:qemuzynqstarter Add a qemu starter for do_testimage
do_testimage will look for a <Machine>starter class and if found for
the given machine will call it's launch() method to launch qemu.
This provides some abstraction for qemu and allows for specific
zynq architecture settings.
The starter provides port and ip details back to the caller and
qemu is setup to route it's serial output to a tcp port so the
do_testimage actions can received the kernel log output.

Signed-off-by: Sipke Vriend <sipke.vriend@xilinx.com>